Fire bullet with mouse
implemented fire bullet system with mouse click

bullet has origin point in center of player

+++ src/com/herkulstudios/entities/BulletShoot.java
@@ -9,13 +9,13 @@ import com.herkulstudios.world.Camera;
 
 public class BulletShoot extends Entity {
 	
-	private int dx;
-	private int dy;
+	private double dx;
+	private double dy;
 	private int life = 100, currentLife = 0;
 	
 	private double spd = 4;
 	
-	public BulletShoot(int x, int y, int width, int height, BufferedImage sprite, int dx, int dy) {
+	public BulletShoot(int x, int y, int width, int height, BufferedImage sprite, double dx, double dy) {
 		super(x, y, width, height, sprite);
 		this.dx = dx;
 		this.dy = dy;
